@@ -21,7 +21,6 @@ import www.raven.jc.dto.UserInfoDTO;
 import www.raven.jc.entity.dto.MessageDTO;
 import www.raven.jc.entity.po.Message;
 import www.raven.jc.entity.po.UserRoom;
-import www.raven.jc.entity.vo.MessageVO;
 import www.raven.jc.event.SaveMsgEvent;
 import www.raven.jc.service.ChatService;
 import www.raven.jc.util.JsonUtil;
@@ -31,7 +30,6 @@ import www.raven.jc.ws.WebsocketService;
 
 import java.util.Date;
 import java.util.List;
-import java.util.stream.Collectors;
 
 /**
  * chat service impl
@@ -42,6 +40,7 @@ import java.util.stream.Collectors;
 @Service
 @Slf4j
 public class ChatServiceImpl implements ChatService {
+
 
     @Autowired
     private MessageDAO messageDAO;
@@ -66,26 +65,25 @@ public class ChatServiceImpl implements ChatService {
         Message realMsg = new Message()
                 .setContent(text)
                 .setTimestamp(new Date(timeStamp))
-                .setSender(user)
+                .setSenderId(user.getUserId())
                 .setType(MessageConstant.ROOM)
                 .setReceiverId(String.valueOf(roomId));
         List<Integer> userIds = userRoomDAO.getBaseMapper().selectList(
                         new QueryWrapper<UserRoom>().eq("room_id", roomId).
                                 eq("status", ApplyStatusConstant.APPLY_STATUS_AGREE)).
-                stream().map(UserRoom::getUserId).collect(Collectors.toList());
+                stream().map(UserRoom::getUserId).toList();
 
         //对离线用户进行离线信息保存
         userIds.forEach(
                 id -> {
                     if (WebsocketService.SESSION_POOL.get(id) == null || !WebsocketService.SESSION_POOL.get(id).isOpen()) {
                         RScoredSortedSet<Object> scoredSortedSet = redissonClient.getScoredSortedSet(OfflineMessagesConstant.PREFIX + id.toString());
-                        MessageVO vo = new MessageVO(realMsg);
-                        log.info("离线消息保存:{}", JsonUtil.objToJson(vo));
-                        scoredSortedSet.add(timeStamp, new MessageVO(realMsg));
+                        log.info("离线消息保存:{}", JsonUtil.objToJson(realMsg));
+                        scoredSortedSet.add(timeStamp, realMsg);
                     }
                 }
         );
-        //异步入历史消息库
+        //异步入历史消息数据库
         streamBridge.send("producer-out-1", MqUtil.createMsg(JsonUtil.objToJson(new SaveMsgEvent().setMessage(realMsg).setType("room")), ImImMqConstant.TAGS_SAVE_HISTORY_MSG));
     }
 
@@ -95,14 +93,14 @@ public class ChatServiceImpl implements ChatService {
         String fixId = MongoUtil.concatenateIds(user.getUserId(), friendId);
         Message realMsg = new Message().setContent(message.getText())
                 .setTimestamp(new Date(message.getTime()))
-                .setSender(user)
+                .setSenderId(user.getUserId())
                 .setType(MessageConstant.FRIEND)
                 .setReceiverId(fixId);
 
         //对离线用户进行离线信息保存
         if (WebsocketService.SESSION_POOL.get(friendId) == null || !WebsocketService.SESSION_POOL.get(friendId).isOpen()) {
             RScoredSortedSet<Object> scoredSortedSet = redissonClient.getScoredSortedSet(OfflineMessagesConstant.PREFIX + user.getUserId());
-            scoredSortedSet.add(message.getTime(), new MessageVO(realMsg));
+            scoredSortedSet.add(message.getTime(), realMsg);
         }
         //异步入历史消息库
         streamBridge.send("producer-out-1", MqUtil.createMsg(JsonUtil.objToJson(new SaveMsgEvent().setMessage(realMsg).setType("friend")), ImImMqConstant.TAGS_SAVE_HISTORY_MSG));
