@@ -2,11 +2,10 @@ package www.raven.jc.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.redisson.api.RScoredSortedSet;
 import org.redisson.api.RedissonClient;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import www.raven.jc.api.UserRpcService;
 import www.raven.jc.constant.OfflineMessagesConstant;
@@ -19,16 +18,18 @@ import www.raven.jc.entity.po.FriendChat;
 import www.raven.jc.entity.po.Message;
 import www.raven.jc.entity.vo.MessageVO;
 import www.raven.jc.entity.vo.UserFriendVO;
+import www.raven.jc.result.RpcResult;
 import www.raven.jc.service.FriendService;
 import www.raven.jc.util.JsonUtil;
 import www.raven.jc.util.MongoUtil;
 import www.raven.jc.util.RequestUtil;
 
-import javax.servlet.http.HttpServletRequest;
+import javax.annotation.Resource;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
@@ -41,15 +42,15 @@ import java.util.stream.Collectors;
 @Service
 @Slf4j
 public class FriendServiceImpl implements FriendService {
-    @Autowired
+    @Resource
     private HttpServletRequest request;
-    @Autowired
+    @Resource
     private FriendChatDAO friendChatDAO;
-    @Autowired
+    @Resource
     private MessageDAO messageDAO;
-    @Autowired
+    @Resource
     private UserRpcService userRpcService;
-    @Autowired
+    @Resource
     private RedissonClient redissonClient;
 
     @Override
@@ -60,7 +61,7 @@ public class FriendServiceImpl implements FriendService {
         if (friends.isEmpty()) {
             return new ArrayList<>();
         }
-        List<Integer> ids = friends.stream().map(UserInfoDTO::getUserId).collect(Collectors.toList());
+        List<Integer> ids = friends.stream().map(UserInfoDTO::getUserId).toList();
         List<String> fixedFriendIds = new ArrayList<>();
         for (Integer friendId : ids) {
             fixedFriendIds.add(MongoUtil.concatenateIds(userId, friendId));
@@ -70,10 +71,12 @@ public class FriendServiceImpl implements FriendService {
         List<String> idsMsg = friendChats.stream().map(FriendChat::getLastMsgId).collect(Collectors.toList());
         //获取好友的最后一条消息
         List<Message> messages = messageDAO.getBaseMapper().selectList(new QueryWrapper<Message>().in("id", idsMsg));
+
+
         //将好友id和好友的最后一条消息id对应起来
         Map<Integer, Message> messageMap = messages.stream()
                 .collect(Collectors.toMap(
-                        message -> message.getSender().getUserId().equals(userId) ? MongoUtil.resolve(message.getReceiverId(), userId) : message.getSender().getUserId(),
+                        message -> message.getSenderId().equals(userId) ? MongoUtil.resolve(message.getReceiverId(), userId) : message.getSenderId(),
                         Function.identity(),
                         (oldValue, newValue) -> newValue
                 ));
@@ -84,7 +87,7 @@ public class FriendServiceImpl implements FriendService {
                     .setFriendName(friend.getUsername())
                     .setFriendProfile(friend.getProfile())
                     .setLastMsg(message == null ? "" : JsonUtil.objToJson(message))
-                    .setLastMsgSender(message == null ? "" : message.getSender().getUserId().equals(userId) ? "我" : friend.getUsername());
+                    .setLastMsgSender(message == null ? "" : message.getSenderId().equals(userId) ? "我" : friend.getUsername());
         }).collect(Collectors.toList());
     }
 
@@ -93,23 +96,31 @@ public class FriendServiceImpl implements FriendService {
         int userId = RequestUtil.getUserId(request);
         String fixId = MongoUtil.concatenateIds(userId, model.getFriendId());
         Page<Message> messagePage = messageDAO.getBaseMapper().selectPage(new Page<>(model.getPage(), model.getSize()), new QueryWrapper<Message>().eq("fix_id", fixId).orderByDesc("timestamp").last("limit 10"));
-        return  messagePage.getRecords().stream().map(MessageVO::new).collect(Collectors.toList());
+        RpcResult<List<UserInfoDTO>> batchInfo = userRpcService.getBatchInfo(List.of(model.getFriendId(), userId));
+        Map<Integer, UserInfoDTO> map = batchInfo.getData().stream().collect(Collectors.toMap(UserInfoDTO::getUserId, Function.identity()));
+        return messagePage.getRecords().stream().map(message -> {
+            UserInfoDTO user = map.get(message.getSenderId());
+            return new MessageVO(message,user);
+        }).collect(Collectors.toList());
     }
 
     @Override
     public List<MessageVO> getLatestFriendMsg(LatestFriendMsgModel model) {
         int userId = RequestUtil.getUserId(request);
-        RScoredSortedSet<MessageVO> scoredSortedSet = redissonClient.getScoredSortedSet(OfflineMessagesConstant.PREFIX + userId);
-        Collection<MessageVO> messages = scoredSortedSet.readAll();
+        RScoredSortedSet<Message> scoredSortedSet = redissonClient.getScoredSortedSet(OfflineMessagesConstant.PREFIX + userId);
+        Collection<Message> messages = scoredSortedSet.readAll();
         //获取所有id=roomId的消息
+        RpcResult<List<UserInfoDTO>> batchInfo = userRpcService.getBatchInfo(List.of(model.getFriendId(), userId));
+        Map<Integer, UserInfoDTO> map = batchInfo.getData().stream().collect(Collectors.toMap(UserInfoDTO::getUserId, Function.identity()));
         List<MessageVO> collect = new ArrayList<>();
-        for (MessageVO message : messages) {
-            if (message.getBelongId().equals(model.getFriendId())) {
-                collect.add(message);
-                //删除已经获取的离线消息
-                scoredSortedSet.remove(message);
+        for (Message message : messages) {
+            if (Objects.equals(message.getReceiverId(), String.valueOf(model.getFriendId()))) {
+                UserInfoDTO user = map.get(message.getSenderId());
+                collect.add(new MessageVO(message,user));
             }
         }
+        //删除已经获取的离线消息
+        scoredSortedSet.removeAll(messages);
         return collect;
     }
 
