@@ -3,8 +3,9 @@ package www.raven.jc.entity.po;
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
-import www.raven.jc.dto.UserInfoDTO;
+import www.raven.jc.model.CommonSerializable;
 
 import java.util.Date;
 
@@ -14,12 +15,13 @@ import java.util.Date;
  * @author 刘家辉
  * @date 2023/11/22
  */
+@EqualsAndHashCode(callSuper = true)
 @Data
 @Accessors(chain = true)
-public class Message {
+public class Message extends CommonSerializable {
     @TableId(value = "id", type = IdType.AUTO)
     private String messageId;
-    private UserInfoDTO sender;
+    private Integer senderId;
     private String content;
     private String type;
     private String receiverId;
