@@ -1,5 +1,8 @@
 package www.raven.jc.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.annotation.Validated;
@@ -15,9 +18,6 @@ import www.raven.jc.result.CommonResult;
 import www.raven.jc.service.UserService;
 import www.raven.jc.util.RequestUtil;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 import java.util.List;
 
 /**
